requirements.txt
tensorflow
opencv-python
numpy
scikit-learn
matplotlib
pyyaml

